build: adapt the component library to iced 0.14
Main upgraded iced 0.13 -> 0.14 and retyped the design tokens from `u16`
to `f32` while this branch was in flight. This carries the feature's own
code across.

API changes absorbed:
- `Widget::on_event` -> `Widget::update`, and `Shell::request_redraw` no
  longer takes a `RedrawRequest`
- `Space::new(w, h)` -> `Space::new().width(w).height(h)`; `with_width` /
  `with_height` are gone
- `checkbox(label, checked)` -> `checkbox(checked).label(label)`
- richer widget statuses: `text_input::Status::Focused` and
  `pick_list::Status::Opened` gained `is_hovered`; `scrollable::Status`
  gained disabled flags

`TypeRole::size` and the glyph primitives now return `f32`, matching the
retyped scale.

**The style-parity fixture was regenerated, and that deserves scrutiny
rather than a shrug.** iced 0.14 changed the shape of its style structs:
the container gained `snap`, the pick-list menu gained a `shadow`, the
scrollable gained an `auto_scroll` indicator, and `Rail.color` became
`Rail.background: Option<Background>`. Before regenerating I normalised
those added fields away and diffed against the previous fixture: **zero**
substantive differences, and every numeric value in the old fixture still
present in the new one. Fields were added; none was altered or removed.
Had any value moved, that would have been an appearance change to explain
rather than a fixture to refresh. The reasoning is recorded in the
snapshot module's own docs so the regeneration stays auditable.

The rustfmt commit was dropped during the rebase — it conflicted with
main's reformatting of the same lines, and re-running the formatter on
the merged result is both simpler and more trustworthy than resolving a
whitespace conflict by hand. Main's retyping also made a batch of
`spacing::X as f32` casts redundant; clippy caught them.

848 passing, 0 failed; fmt and clippy clean. All three of the feature's
gates hold: the boundary still reads 0/0/0, the API ratchet is unchanged,
and the style snapshot matches.

+++ crates/micold-client/src/ui/sidebar.rs
@@ -9,9 +9,9 @@ use crate::ui::material::{
 };
 use iced::widget::{column, container, mouse_area, row, Space};
 use iced::{Alignment, Element, Length};
-use micold_core::tokens::{self, sidebar, spacing, Rgb, Roles};
 use micold_core::naming::Tag;
 use micold_core::session::{SessionLifecycle, SessionLocation};
+use micold_core::tokens::{self, sidebar, spacing, Rgb, Roles};
 use micold_core::worktree::WorktreeStatus;
 
 /// Width of the draggable resize handle between the sidebar and the main area.
@@ -152,10 +152,10 @@ pub fn view<'a>(
     let content = column![header, filter_accordion, body]
         .spacing(spacing::SM)
         .padding(iced::Padding {
-            top: spacing::SM as f32,
-            bottom: spacing::SM as f32,
-            left: spacing::XS as f32,
-            right: spacing::XS as f32,
+            top: spacing::SM,
+            bottom: spacing::SM,
+            left: spacing::XS,
+            right: spacing::XS,
         })
         .width(Length::Fixed(width))
         .height(Length::Fill);
